chore: 댓글 입력 뷰 분리
